feat: move requisicão de usuário DE AboutMe para Searchbar, remove prop drilling e consome dados da store para evitar requisição duplicada
