+++ src/components/SearchBar/index.tsx
@@ -1,19 +1,53 @@
 import React from 'react'
-import { getUser } from 'hooks'
 import { MdMenu } from 'react-icons/md'
+import { useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
+import { AppDispatch } from 'redux/store'
 import { useSidebarCollapse } from 'contexts'
 import { useNavigate } from 'react-router-dom'
-import { Avatar, Flex, HStack, IconButton, Input } from '@chakra-ui/react'
+import { getUser, getUserId, useMedia } from 'hooks'
+import { fetchAboutMe, getAboutUser } from '../../redux/reducers/user'
+import {
+  Avatar,
+  Flex,
+  HStack,
+  IconButton,
+  Input,
+  useToast,
+} from '@chakra-ui/react'
 
 export const SearchBar = () => {
+  const toast = useToast()
   const navigate = useNavigate()
 
-  const [user, setUser] = React.useState<string>()
+  const dispatch = useDispatch<AppDispatch>()
+
+  // const [user, setUser] = React.useState<string>()
+  const { isMobileOrTablet } = useMedia()
 
   const { collapse, setCollapse } = useSidebarCollapse()
 
+  const user = useSelector(getAboutUser)
+
+  const fetchUser = React.useCallback(async () => {
+    const id = getUserId()
+    try {
+      // const response = await UserApi.findMe(id!)
+
+      dispatch(fetchAboutMe(id!))
+    } catch (e) {
+      toast({
+        description: 'Erro ao carregar dados do usuário',
+        status: 'error',
+        containerStyle: { color: 'white' },
+        position: isMobileOrTablet ? 'top' : 'bottom-right',
+        isClosable: true,
+      })
+    }
+  }, [])
+
   React.useEffect(() => {
-    setUser(getUser() ?? '')
+    fetchUser()
   }, [])
 
   return (
@@ -49,8 +83,9 @@ export const SearchBar = () => {
           }}
           focusBorderColor="transparent"
         />
+
         <Avatar
-          name={user}
+          name={user.name}
           w="1.75rem"
           h="1.75rem"
           color="white"
